# services/fleet-svc/openapi.yaml
openapi: 3.1.0
info:
  title: Sonalit Fleet Service
  version: 4.0.0
servers:
  - url: http://localhost:4002
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /v4/vehicles:
    get:
      summary: List vehicles
      operationId: listVehicles
      tags: [Vehicles]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, maintenance, retired]
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Paginated list of vehicles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a vehicle
      operationId: createVehicle
      tags: [Vehicles]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plate, make, model, year, fuel_type]
              properties:
                plate: { type: string, minLength: 1, maxLength: 20 }
                make: { type: string, minLength: 1, maxLength: 100 }
                model: { type: string, minLength: 1, maxLength: 100 }
                year: { type: integer, minimum: 1900, maximum: 2100 }
                vin: { type: string, maxLength: 17 }
                fuel_type:
                  type: string
                  enum: [petrol, diesel, electric, hybrid, cng, lpg]
                status:
                  type: string
                  enum: [active, inactive, maintenance, retired]
                  default: active
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /v4/vehicles/{id}:
    get:
      summary: Get a vehicle
      operationId: getVehicle
      tags: [Vehicles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Vehicle record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      summary: Update a vehicle
      operationId: updateVehicle
      tags: [Vehicles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plate: { type: string, minLength: 1, maxLength: 20 }
                make: { type: string, minLength: 1, maxLength: 100 }
                model: { type: string, minLength: 1, maxLength: 100 }
                year: { type: integer, minimum: 1900, maximum: 2100 }
                vin: { type: string, maxLength: 17 }
                fuel_type:
                  type: string
                  enum: [petrol, diesel, electric, hybrid, cng, lpg]
                status:
                  type: string
                  enum: [active, inactive, maintenance, retired]
      responses:
        '200':
          description: Updated vehicle record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Soft-delete a vehicle
      operationId: deleteVehicle
      tags: [Vehicles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Vehicle deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # Drivers
  # ─────────────────────────────────────────────
...
  /v4/shipments:
    get:
      summary: List shipments
      operationId: listShipments
      tags: [Shipments]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_transit, delivered, cancelled, failed]
        - name: vehicle_id
          in: query
          schema: { type: string, format: uuid }
        - name: driver_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Paginated list of shipments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Shipment'
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a shipment
      operationId: createShipment
      tags: [Shipments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination, scheduled_at]
              properties:
                origin: { type: string, minLength: 1, maxLength: 500 }
                destination: { type: string, minLength: 1, maxLength: 500 }
                cargo_description: { type: string, maxLength: 2000 }
                status:
                  type: string
                  enum: [pending, assigned, in_transit, delivered, cancelled, failed]
                  default: pending
                vehicle_id: { type: string, format: uuid }
                driver_id: { type: string, format: uuid }
                scheduled_at: { type: string, format: date-time }
                delivered_at: { type: string, format: date-time }
      responses:
        '201':
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v4/shipments/{id}:
    get:
      summary: Get a shipment
      operationId: getShipment
      tags: [Shipments]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Shipment record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      summary: Update a shipment
      operationId: updateShipment
      tags: [Shipments]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                origin: { type: string, minLength: 1, maxLength: 500 }
                destination: { type: string, minLength: 1, maxLength: 500 }
                cargo_description: { type: string, maxLength: 2000 }
                status:
                  type: string
                  enum: [pending, assigned, in_transit, delivered, cancelled, failed]
                vehicle_id: { type: string, format: uuid }
                driver_id: { type: string, format: uuid }
                scheduled_at: { type: string, format: date-time }
                delivered_at: { type: string, format: date-time }
      responses:
        '200':
          description: Updated shipment record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Soft-delete a shipment
      operationId: deleteShipment
      tags: [Shipments]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Shipment deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # Messages
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
  schemas:
    Vehicle:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        plate: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        vin: { type: string }
        fuel_type: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
...
    Shipment:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        shipment_number: { type: string }
        origin: { type: string }
        destination: { type: string }
        cargo_description: { type: string, nullable: true }
        status: { type: string }
        vehicle_id: { type: string, format: uuid, nullable: true }
        driver_id: { type: string, format: uuid, nullable: true }
        scheduled_at: { type: string, format: date-time }
        delivered_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
...